requirements.txt
# Google Cloud SDK
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# CLI
click==8.1.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0
